requirements.txt
authlib
Flask
google-cloud-secret-manager
gunicorn
python-dotenv
requests